Clean up FdbServer API
