+++ src/FdbServer/IRunningFdbServer.cs
@@ -2,7 +2,7 @@
 
 namespace FdbServer
 {
-    public interface IRunningFdbServer : IDisposable
+    public interface IRunningFdbServer : IFdbServer, IDisposable
     {
         /// <summary>
         /// Initializes the server, required on newly created servers.
