+++ src/FdbServer/IStoppedFdbServer.cs
@@ -1,7 +1,12 @@
 ﻿namespace FdbServer
 {
-    public interface IStoppedFdbServer
+    public interface IStoppedFdbServer : IFdbServer
     {
+        /// <summary>
+        /// Starts the server.
+        /// </summary>
+        IRunningFdbServer Start();
+
         /// <summary>
         /// Destroys the server, deleting all files created.
         /// </summary>
